reworked movement to use center of tile and center of object, pathfinding still needs to be changed

+++ Digger/AimCommand.cpp
@@ -1,6 +1,6 @@
 #include "AimCommand.h"
 
-dae::AimCommand::AimCommand(glm::vec2 direction, ShootingComponent* shootComp)
+dae::AimCommand::AimCommand(glm::ivec2 direction, ShootingComponent* shootComp)
 	:m_Direction(direction), m_pShootComp(shootComp)
 {
