+++ Minigin/Transform.cpp
@@ -1,6 +1,6 @@
 #include "Transform.h"
 
-void dae::Transform::SetPosition(const float x, const float y, const float z)
+void dae::Transform::SetPosition(const int x, const int y, const int z)
 {
 	m_position.x = x;
 	m_position.y = y;
