+++ Minigin/Transform.h
@@ -7,8 +7,8 @@ namespace dae
 	{
 	public:
 		const glm::vec3& GetPosition() const { return m_position; }
-		void SetPosition(float x, float y, float z);
+		void SetPosition(int x, int y, int z);
 	private:
-		glm::vec3 m_position;
+		glm::ivec3 m_position;
 	};
 }
